[IMP] General refactor

+++ app/src/main/java/ocr/document/tardo/documentocr/activities/LoginActivity.java
@@ -12,6 +12,8 @@ import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.os.AsyncTask;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.design.widget.TextInputEditText;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.content.ContextCompat;
 import android.util.Base64;
@@ -20,7 +22,6 @@ import android.view.View.OnClickListener;
 import android.widget.ArrayAdapter;
 import android.widget.AutoCompleteTextView;
 import android.widget.Button;
-import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.ProgressBar;
 import android.widget.TextView;
@@ -34,14 +35,13 @@ import com.googlecode.tesseract.android.TessBaseAPI;
 import org.json.JSONArray;
 import org.json.JSONException;
 
-import java.io.UnsupportedEncodingException;
 import java.net.MalformedURLException;
 import java.nio.charset.StandardCharsets;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
 import java.security.spec.InvalidKeySpecException;
 import java.util.ArrayList;
-import java.util.List;
+import java.util.Objects;
 
 import javax.crypto.BadPaddingException;
 import javax.crypto.Cipher;
@@ -63,15 +63,15 @@ public class LoginActivity extends AccountAuthenticatorActivity implements OnCli
     private final int REQUEST_PERMISSION_WRITE_EXTERNAL_STORAGE = 1;
 
     public static String PARAM_LOGOUT = "logout";
-    private EditText mEditHost;
-    private AutoCompleteTextView mEditDBName;
-    private EditText mEditLogin;
-    private EditText mEditPass;
+    public TextInputEditText mEditHost;
+    public AutoCompleteTextView mEditDBName;
+    public TextInputEditText mEditLogin;
+    public TextInputEditText mEditPass;
     private TextView mTextError;
     private Button mBtnLoginIn;
     private ProgressBar mProgressBar;
     private ImageView mLogo;
-    private SharedPreferences mSettings;
+    public SharedPreferences mSettings;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -107,7 +107,7 @@ public class LoginActivity extends AccountAuthenticatorActivity implements OnCli
         mBtnLoginIn.setOnClickListener(this);
         mEditDBName.setOnClickListener(this);
 
-        new GetDBTask().execute(mEditHost.getText().toString());
+        new GetDBTask().execute(Objects.requireNonNull(mEditHost.getText()).toString());
 
         // Comprobar que la app tiene permisos suficientes para funcionar correctamente
         if (ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED) {
@@ -123,20 +123,19 @@ public class LoginActivity extends AccountAuthenticatorActivity implements OnCli
     }
 
     @Override
-    public void onRequestPermissionsResult(int requestCode, String permissions[], int[] grantResults) {
-        switch (requestCode) {
-            case REQUEST_PERMISSION_WRITE_EXTERNAL_STORAGE:
-                if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                    initTesseract();
-                    // Ignorar login si ya se logeo correctamente
-                    if (mSettings.getInt("UserID", -1) > 0)
-                        initApp();
-                    Toast.makeText(LoginActivity.this, "Permission Granted!", Toast.LENGTH_SHORT).show();
-                } else {
-                    // TODO: No ser tan duro y mostrar algun mensaje de error
-                    android.os.Process.killProcess(android.os.Process.myPid());
-                    System.exit(0);
-                }
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        if (requestCode == REQUEST_PERMISSION_WRITE_EXTERNAL_STORAGE) {
+            if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
+                initTesseract();
+                // Ignorar login si ya se logeo correctamente
+                if (mSettings.getInt("UserID", -1) > 0)
+                    initApp();
+                Toast.makeText(LoginActivity.this, "Permission Granted!", Toast.LENGTH_SHORT).show();
+            } else {
+                // TODO: No ser tan duro y mostrar algun mensaje de error
+                android.os.Process.killProcess(android.os.Process.myPid());
+                System.exit(0);
+            }
         }
     }
 
@@ -146,9 +145,10 @@ public class LoginActivity extends AccountAuthenticatorActivity implements OnCli
         if (v.getId() == R.id.btnLogin)
         {
             new LoginTask().execute(
-                    mEditHost.getText().toString(),
-                    mEditDBName.getText().toString(), mEditLogin.getText().toString(),
-                    mEditPass.getText().toString());
+                    Objects.requireNonNull(mEditHost.getText()).toString(),
+                    mEditDBName.getText().toString(),
+                    Objects.requireNonNull(mEditLogin.getText()).toString(),
+                    Objects.requireNonNull(mEditPass.getText()).toString());
             mTextError.setVisibility(View.INVISIBLE);
             hideControls(Boolean.TRUE);
         } else if (v.getId() == R.id.editDBName)
@@ -160,7 +160,7 @@ public class LoginActivity extends AccountAuthenticatorActivity implements OnCli
     @Override
     public void onFocusChange(View v, boolean b) {
         if (!b) {
-            new GetDBTask().execute(mEditHost.getText().toString());
+            new GetDBTask().execute(Objects.requireNonNull(mEditHost.getText()).toString());
         }
     }
 
@@ -192,27 +192,15 @@ public class LoginActivity extends AccountAuthenticatorActivity implements OnCli
         String passwd = "";
         try {
             byte[] encrypedPwdBytes = Base64.decode(mSettings.getString("Pass", ""), Base64.NO_WRAP);
-            DESKeySpec keySpec = new DESKeySpec(mSettings.getString("rn", "").getBytes());
+            DESKeySpec keySpec = new DESKeySpec(Objects.requireNonNull(mSettings.getString("rn", "")).getBytes());
             SecretKeyFactory keyFactory = SecretKeyFactory.getInstance("DES");
             SecretKey key = keyFactory.generateSecret(keySpec);
-            Cipher cipher = Cipher.getInstance("DES");
+            Cipher cipher = Cipher.getInstance("DES/OFB32/PKCS5Padding");
             cipher.init(Cipher.DECRYPT_MODE, key);
             byte[] plainTextPwdBytes = (cipher.doFinal(encrypedPwdBytes));
             passwd = new String(plainTextPwdBytes, StandardCharsets.UTF_8);
-        } catch (NoSuchAlgorithmException e) {
-            e.printStackTrace();
-        } catch (InvalidKeyException e) {
-            e.printStackTrace();
-        } catch (NoSuchPaddingException e) {
-            e.printStackTrace();
-        } catch (BadPaddingException e) {
-            e.printStackTrace();
-        } catch (InvalidKeySpecException e) {
-            e.printStackTrace();
-        } catch (IllegalBlockSizeException e) {
-            e.printStackTrace();
-        } catch (IllegalArgumentException e) {
-            e.printStackTrace();
+        } catch (NoSuchAlgorithmException | InvalidKeyException | NoSuchPaddingException | BadPaddingException | InvalidKeySpecException | IllegalBlockSizeException | IllegalArgumentException e) {
+            e.printStackTrace(); // TODO: It's an error, don't hide & forget it ¬¬
         }
 
         try {
@@ -225,7 +213,12 @@ public class LoginActivity extends AccountAuthenticatorActivity implements OnCli
             return;
         }
 
-        new CheckAndInitAppTask().execute();
+        // Init
+        SharedPreferences.Editor editor = mSettings.edit();
+        editor.putBoolean("HasHotelL10N", Boolean.TRUE); // FIXME: Hardcoded value
+        editor.apply();
+        startActivity(new Intent(this, ReadModeActivity.class));
+        finish();
     }
 
     private void hideControls(Boolean state) {
@@ -249,7 +242,6 @@ public class LoginActivity extends AccountAuthenticatorActivity implements OnCli
         private Exception mException;
         private Integer mUID;
 
-
         protected Boolean doInBackground(String... params) {
             mException = null;
             try {
@@ -270,32 +262,20 @@ public class LoginActivity extends AccountAuthenticatorActivity implements OnCli
             if (res) {
                 SharedPreferences.Editor editor = mSettings.edit();
                 editor.putInt("UserID", mUID);
-                editor.putString("Host", mEditHost.getText().toString());
+                editor.putString("Host", Objects.requireNonNull(mEditHost.getText()).toString());
                 editor.putString("DBName", mEditDBName.getText().toString());
 
                 try {
-                    byte[] cleartext = mEditPass.getText().toString().getBytes("UTF8");
-                    DESKeySpec keySpec = new DESKeySpec(mSettings.getString("rn", "").getBytes());
+                    byte[] cleartext = Objects.requireNonNull(mEditPass.getText()).toString().getBytes(StandardCharsets.UTF_8);
+                    DESKeySpec keySpec = new DESKeySpec(Objects.requireNonNull(mSettings.getString("rn", "")).getBytes());
                     SecretKeyFactory keyFactory = SecretKeyFactory.getInstance("DES");
                     SecretKey key = keyFactory.generateSecret(keySpec);
-                    Cipher cipher = Cipher.getInstance("DES");
+                    Cipher cipher = Cipher.getInstance("DES/OFB32/PKCS5Padding");
                     cipher.init(Cipher.ENCRYPT_MODE, key);
                     final String encPasswdBase64 = Base64.encodeToString(cipher.doFinal(cleartext), Base64.NO_WRAP);
                     editor.putString("Pass", encPasswdBase64);
-                } catch (InvalidKeyException e) {
-                    e.printStackTrace();
-                } catch (UnsupportedEncodingException e) {
-                    e.printStackTrace();
-                } catch (NoSuchAlgorithmException e) {
-                    e.printStackTrace();
-                } catch (InvalidKeySpecException e) {
-                    e.printStackTrace();
-                } catch (NoSuchPaddingException e) {
-                    e.printStackTrace();
-                } catch (IllegalBlockSizeException e) {
-                    e.printStackTrace();
-                } catch (BadPaddingException e) {
-                    e.printStackTrace();
+                } catch (InvalidKeyException | NoSuchAlgorithmException | InvalidKeySpecException | NoSuchPaddingException | IllegalBlockSizeException | BadPaddingException e) {
+                    e.printStackTrace(); // TODO: It's an error, don't hide & forget it ¬¬
                 }
 
                 editor.apply();
@@ -303,35 +283,9 @@ public class LoginActivity extends AccountAuthenticatorActivity implements OnCli
                 initApp();
             }
             else {
-               showErrorMessage(
+                showErrorMessage(
                        (mException != null)?mException.getMessage():getResources().getString(R.string.login_invalid));
             }
-        }
-    }
-
-    private class CheckAndInitAppTask extends AsyncTask<String, Void, Boolean> {
-        private Boolean mHasHotelL10N = false;
-
-
-        protected Boolean doInBackground(String... params) {
-            final JSONRPCClientOdoo OdooClient = ((AppMain)getApplication()).OdooClient();
-            // Check if enable support for hotel_l10n_es
-            try {
-                mHasHotelL10N = OdooClient.callCount(
-                        "ir.module.module",
-                        "[['state', '=', 'installed'], ['name', '=', 'hotel_l10n_es']]") > 0;
-            } catch (OdooSearchException e) {
-                e.printStackTrace();
-            }
-            return Boolean.TRUE;
-        }
-
-        protected void onPostExecute(Boolean res) {
-            SharedPreferences.Editor editor = mSettings.edit();
-            editor.putBoolean("HasHotelL10N", res && mHasHotelL10N);
-            editor.apply();
-            startActivity(new Intent(LoginActivity.this, ReadModeActivity.class));
-            finish();
         }
     }
 
@@ -344,12 +298,11 @@ public class LoginActivity extends AccountAuthenticatorActivity implements OnCli
             try {
                 OdooClient = new JSONRPCClientOdoo(params[0]);
                 mDBList = OdooClient.getDBList();
-            } catch (OdooSearchException e) {
-                e.printStackTrace();
-            } catch (MalformedURLException e) {
-                e.printStackTrace();
+                return Boolean.TRUE;
+            } catch (OdooSearchException | MalformedURLException e) {
+                e.printStackTrace(); // TODO: It's an error, don't hide & forget it ¬¬
             }
-            return Boolean.TRUE;
+            return Boolean.FALSE;
         }
 
         protected void onPostExecute(Boolean res) {
